+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Next
+*2021-06-05*
+- Added CLI option `--full` to log a full list of parsed files.
+
 ## 1.0.1
 *2021-06-05*
 - Fixed command-line usage not working.